tidy(ui): organise canvas tool classes
